Rip the top-level xRI path parsing implementation out of ... ... the body of the top-level xRI parser and push it down into the new onm-xri-parse-path module. Note that one of the tasks on this little sub-adventure is to add support and tests for both abstract and concrete resource identifers. This is pretty abstract and introduces an evolution in my own terminology...
onm.Address is a runtime resource identifier.
xRI's are strings that are _partial_ serializations of onm.Address.
You cannot construct an onm.Address from an xRI without an onm.Model.
xRI paths assert a walk through a model's namespace graph.
When a path asserts a walk through an extension point namespace, the asserting token is special.
The special token indicates one of several possible embedded subrequests related to the disposition of the addressed data component.
If this token is equal to '' (zero-length string), '-', or the namespace model's declared 'jsonTag', then the subrequest is to identify the data component abstractly. Otherwise, the subrequest is to identify a specific instance of the data component namespace resource by name.

onm's new 'resolver' algorithm combines (a) a resource locator algorithm that operates on onm.Address and produces onm.Namespace (b) a data transformation infrastructure (currently leveraged for namespace property initialization and data overwrite. later to be extended in ways...).

An "abstract" path xRI (i.e. a path that refers to one or more data components abstractly) can never be "opened" by the onm resolver (it doesn't exist). It can however be "created". It's critical to understand that the process of creating a resource from an abstract resource identifier generates a new resource identifier that is "concrete". This concrete resource identifier is held along with the created resource by onm.Namespace.

A "conrete" path xRI (i.e. a path that refers to all referenced data components literally by name) may be resolvable via "open" if the specified resource actually exists. Otherwise, it should be resolvable via "create" with the supposition that the resultant onm.Namespace is bound to the same initial conrete path.